prototype multi-part table model
This is a real-sized piece of furniture! It includes a base, a top, and
a leg piece. You can print multiple 6-inch legs and screw them together
to make the table whatever height you want. By default, the table top is
a circle with a radius of 7 inches, which is enough to hold a phone or a
glass of water.