move lambert's law out of BRDF!
The BRDF shouldn't include the cosine of the incoming angle. That cosine
term is implicit in the density of rays that actually hits the light
source. The only time it should be added implicitly is by point lights,
which needed a coefficient fix anyway due to the hemisphere BRDF density
change.

This changes the rendering of all materials (Phong and Lambert) to be
more correct during ray tracing. There is no effect for ray casting.