smooth join API
This API makes it possible to join together multiple SDFs with rounded,
smooth intersections.

I attempted to apply this API to the corgi example, but the corgi is
very over-engineered to its current smoothing technique. As a result,
switching to an SDF smoothing technique revealed a lot of pecularities
in the parts of the model (e.g. squared-off ears, huge tail nub).

Fixes #24