deformation: use sparse cholesky, no CG
This improves speed significantly on the example, and it should improve
speed more and more as more iterations are needed to converge.

Currently, the sparse cholesky implementation is not as efficient as it
could be, since it uses RCM rather than a more advanced re-ordering
algorithm. However, this approach is simple and self-contained.