model2d: fix export for unclosed paths
When we export a curve or some other unclosed path, it would generate several smaller paths, since it would start at some random point and scan until the end of the path, and then repeat. This was not only semantically incorrect, but it looked wrong because of my use of <polygon> when I should use <polyline> in the SVG file. I have now changed the way paths are searched over (now it's bidirectional) and I've added the use of polyline for unclosed paths.