test to show bad complexity of MeshToHierarchy
In its current form, MeshToHierarchy() is N^2 in the number of submeshes, since after each sub-mesh we re-iterate over the entire mesh to find the "lowest" coordinate. Additionally, it has to iterate over all previously created solids to find where in the hierarchy a given mesh goes, although this seems to require much less compute in practice.